requirements_real.txt
pandas>=2.0.0
openpyxl>=3.1.0
tqdm>=4.65.0
yfinance>=0.2.28
requests>=2.31.0
google-generativeai>=0.3.0
